requirements.txt
flask==2.3.3
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78
matplotlib==3.7.2
kaggle==1.5.16
